requirements.txt
# Core dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
paho-mqtt>=1.6.0

# Web dashboard
flask>=2.3.0
plotly>=5.17.0
pandas>=2.0.0

# Utilities
numpy>=1.24.0